refactor(gui): split controls and remove effect renders
